Now Commit working code

+++ cypress/support/fakerAdditionallease.js
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker';
 
 export const generateFakeleasedata = () => {
-  const area = faker.number();
+  const area = faker.number.int();
   const designation = faker.person.jobTitle();
   const comments = faker.internet.displayName();
   return {
